fix: 패키지 구조 수정
- 전체적인 디렉토리 구조 수정
- 토큰 검증 앤드포인트 생성
