+++ build.gradle
@@ -14,6 +14,14 @@ allprojects {
     }
 }
 
+tasks.named('bootJar') {
+    enabled = false
+}
+
+tasks.named('jar') {
+    enabled = false
+}
+
 configurations {
     compileOnly {
         extendsFrom annotationProcessor
