+++ monglife-authorization-auth-api/src/main/java/com/monglife/authorization/auth/business/vo/ReissueVo.java
@@ -1,9 +1,9 @@
-package com.monglife.authorization.auth.dto.res;
+package com.monglife.authorization.auth.business.vo;
 
 import lombok.Builder;
 
 @Builder
-public record ReissueResDto(
+public record ReissueVo(
         String accessToken,
         String refreshToken
 ) {
