+++ monglife-authorization-auth-api/src/main/java/com/monglife/authorization/auth/business/vo/ValidationAccessTokenVo.java
@@ -1,0 +1,9 @@
+package com.monglife.authorization.auth.business.vo;
+
+import lombok.Builder;
+
+@Builder
+public record ValidationAccessTokenVo(
+        String accessToken
+) {
+}
