+++ monglife-authorization-auth-api/src/main/java/com/monglife/authorization/auth/presentation/dto/req/LoginReqDto.java
@@ -1,4 +1,4 @@
-package com.monglife.authorization.auth.dto.req;
+package com.monglife.authorization.auth.presentation.dto.req;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
