+++ monglife-authorization-auth-api/src/main/java/com/monglife/authorization/auth/presentation/dto/req/LogoutReqDto.java
@@ -1,9 +1,9 @@
-package com.monglife.authorization.auth.dto.req;
+package com.monglife.authorization.auth.presentation.dto.req;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 
-public record ReissueReqDto(
+public record LogoutReqDto(
         @NotEmpty
         @NotBlank
         String refreshToken
