+++ monglife-authorization-auth-api/src/main/java/com/monglife/authorization/auth/presentation/dto/res/LoginResDto.java
@@ -1,4 +1,4 @@
-package com.monglife.authorization.auth.dto.res;
+package com.monglife.authorization.auth.presentation.dto.res;
 
 import lombok.Builder;
