+++ monglife-authorization-auth-api/src/main/java/com/monglife/authorization/auth/presentation/dto/res/ValidationAccessTokenResDto.java
@@ -1,0 +1,9 @@
+package com.monglife.authorization.auth.presentation.dto.res;
+
+import lombok.Builder;
+
+@Builder
+public record ValidationAccessTokenResDto(
+        String accessToken
+) {
+}
